% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_viewer.R
\name{launch_viewer}
\alias{launch_viewer}
\title{Start the nwmViewer shiny interface}
\usage{
launch_viewer()
}
\description{
Start the nwmViewer shiny interface
}
\examples{
# launch_viewer()
}
\keyword{GUI}
